+++ src/Components/Shop/ShopContainer.js
@@ -1,5 +1,0 @@
-import React from "react";
-export {
-  CartContainer as ShopContainer,
-  store
-} from "../Cart/CartContainer";