+++ src/hooks/usePelcro/index.js
@@ -1,42 +1,55 @@
 import createHook from "zustand";
 import createStore from "zustand/vanilla";
 import { mountStoreDevtool } from "simple-zustand-devtools";
+import { persist } from "zustand/middleware";
 import { PelcroActions } from "./pelcroActions";
 import { PelcroCallbacks } from "./pelcroCallbacks";
 
+export const initialState = {
+  // View
+  view: null,
+
+  // Plans
+  product: null,
+  plan: null,
+  isGift: false,
+  isRenewingGift: false,
+  giftCode: "",
+  subscriptionIdToRenew: null,
+
+  // E-commerce
+  cartItems: [],
+
+  // User
+  isAuthenticated: () => window.Pelcro.user.isAuthenticated(),
+  selectedAddressId: null,
+  addressIdToEdit: null,
+
+  // internal meta-state
+  meta: {
+    orderedItems: []
+  }
+};
+
 const createPelcroStore = () =>
-  createStore((set, get) => {
-    const actions = new PelcroActions(set, get);
-    const eventCallbacks = new PelcroCallbacks();
-    return {
-      // Store setter
-      set,
-
-      // View
-      view: null,
-
-      // Plans
-      product: null,
-      plan: null,
-      isGift: false,
-      isRenewingGift: false,
-      giftCode: "",
-      subscriptionIdToRenew: null,
-
-      // E-commerce
-      order: null,
-
-      // User
-      isAuthenticated: () => window.Pelcro.user.isAuthenticated(),
-      selectedAddressId: null,
-      addressIdToEdit: null,
-
-      // State actions
-      ...actions,
-      // Callbacks
-      ...eventCallbacks
-    };
-  });
+  createStore(
+    persist(
+      (set, get) => {
+        const actions = new PelcroActions(set, get);
+        const eventCallbacks = new PelcroCallbacks();
+        return {
+          // Store setter
+          set,
+          ...initialState,
+          // State actions
+          ...actions,
+          // Callbacks
+          ...eventCallbacks
+        };
+      },
+      { name: "pelcro-store", whitelist: ["cartItems"] }
+    )
+  );
 
 const createPelcroHook = (store) => {
   // initiate hook with zustand creator