+++ webpack.config.js
@@ -26,6 +26,8 @@ module.exports = {
           presets: ["@babel/preset-env", "@babel/preset-react"],
           plugins: [
             "@babel/plugin-proposal-class-properties",
+            "@babel/plugin-proposal-optional-chaining",
+            "@babel/plugin-proposal-nullish-coalescing-operator",
             "react-refresh/babel"
           ]
         }